backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
python-docx>=1.1,<2.0
python-pptx>=0.6.23,<1.0
pypdf>=4.2,<6.0
openai>=1.0,<2.0
python-multipart>=0.0.9,<1.0
